refactor(self_relaunch_semaphore): split tests.rs (#1266)
tests.rs 443 -> 231 LOC; new tests_extra.rs (~235 LOC).
